Enable syncing without composer update
It was pointed out that with the current setup if the initial
CI is merged, it could sometimes fail when composer libraries
of extensions get new releases that fit within the version
restrictions that are specified.

So, don't delete the composer.lock file, and only run composer
update when a user specifically requests it when building this repo
(Which would be done when 1) looking for updates or 2) adding new
extensions etc)

+++ wbstack/sync/04-docker-composer.sh
@@ -1,23 +1,26 @@
 #!/usr/bin/env sh
 
-echo "Performing needed composer installations"
-
 SCRIPT_COMPOSER_CACHE=${COMPOSER_CACHE_DIR:-$HOME/.cache/composer}
 
 mkdir -p ${COMPOSER_CACHE_DIR:-$HOME/.cache/composer}
 
 # composer install
+echo "Performing needed composer installations"
 docker run --rm -it -u $(id -u ${USER}):$(id -g ${USER}) -v $PWD:/app \
   --volume $SCRIPT_COMPOSER_CACHE:/tmp/cache \
   --entrypoint composer -w /app \
   docker-registry.wikimedia.org/releng/composer-package-php74:0.3.0-s7 install --no-dev --no-progress --optimize-autoloader
 
-# composer update
-docker run --rm -it -u $(id -u ${USER}):$(id -g ${USER}) -v $PWD:/app \
-  --volume $SCRIPT_COMPOSER_CACHE:/tmp/cache \
-  --entrypoint composer -w /app \
-  docker-registry.wikimedia.org/releng/composer-package-php74:0.3.0-s7 update --no-dev --no-progress --optimize-autoloader
-
+# composer update (When ALSO_COMPOSER_UPDATE = 1)
+if [ "${ALSO_COMPOSER_UPDATE}" = "1" ]; then
+    echo "Performing composer update"
+    docker run --rm -it -u $(id -u ${USER}):$(id -g ${USER}) -v $PWD:/app \
+      --volume $SCRIPT_COMPOSER_CACHE:/tmp/cache \
+      --entrypoint composer -w /app \
+      docker-registry.wikimedia.org/releng/composer-package-php74:0.3.0-s7 update --no-dev --no-progress --optimize-autoloader
+else
+    echo "SKIPPING: composer update (As you didn't request it)"
+fi;
 
 # Sometimes composer git clones things rather than using zips.
 # so make sure to remove the .git directories so that things get committed correctly.
